Add Jest testing infrastructure
