+++ README.md
@@ -2,5 +2,15 @@
 
 # Wooflejelly
 
-Wooflejelly is, at present, an aspirational attempt to recreate 
-the [SPDX-License-Diff](https://github.com/spdx/spdx-license-diff) browser plugin (apparently not actively maintained as of 2025) as an Elm application.  
+Wooflejelly is, at present, an aspirational attempt to recreate
+the [SPDX-License-Diff](https://github.com/spdx/spdx-license-diff) browser plugin (apparently not actively maintained as of 2025) as an Elm application.
+
+## Running tests
+
+Install dependencies once using `npm install`, then run the suite with:
+
+```bash
+npm test
+```
+
+This uses Jest together with `ts-jest` to execute the TypeScript tests under the `tests/` directory.
