+++ src/popup.ts
@@ -4,7 +4,7 @@ console.log("popup.ts loaded");
 
 console.log("popup.ts top-level code running");
 
-
+import { computeSimilarity } from './similarity';
 
 console.log("typeof diff_match_patch:", typeof diff_match_patch);
 
@@ -43,31 +43,6 @@ async function loadSpdxLicenses(): Promise<LicenseText[]> {
 type Diff = [number, string];
 
 
-function computeSimilarity(a: string, b: string): number {
-  console.log("computeSimilarity a:", a.slice(0, 30));
-  console.log("computeSimilarity b:", b.slice(0, 30));
-  console.log("types:", typeof a, typeof b);
-
-  const dmp = new diff_match_patch();
-  const diff = dmp.diff_main(a, b);
-  dmp.diff_cleanupSemantic(diff);
-  console.log("raw diff_main output:", diff);
-
-  // Handle diff_match_patch.Diff instances instead of [op, text] arrays
-  const equalParts = diff.filter((entry: any) => entry?.[0] === 0);
-
-  console.log("filtered equal parts:", equalParts);
-
-  const equalLength = equalParts.reduce((sum: number, entry: any) => {
-    return sum + (entry?.[1]?.length || 0);
-  }, 0);
-
-  const totalLength = a.length + b.length;
-  const similarity = totalLength === 0 ? 1 : (2 * equalLength) / totalLength;
-
-  console.log("similarity:", similarity.toFixed(4));
-  return similarity;
-}
