+++ tsconfig.json
@@ -3,7 +3,7 @@
     "target": "ES6",
     "module": "ES6",
     "lib": ["DOM", "ES2020"],
-    "types": ["chrome"],
+    "types": ["chrome", "jest", "node"],
     "strict": true,
     "moduleResolution": "node",
     "outDir": "dist", 
@@ -11,7 +11,7 @@
     "checkJs": false,
     "esModuleInterop": true
   },
-  "include": ["src"],
+  "include": ["src", "tests"],
   "exclude": [
     "src/elm.js",
     "src/diff_match_patch.js",
